+++ engineer/models.py
@@ -70,6 +70,8 @@ class Post(object):
         if not hasattr(self, 'content_preprocessed'):
             self.content_preprocessed = self.content_raw
 
+        self._content_finalized = self.content_raw
+
         # Handle any preprocessor plugins
         for plugin in PostProcessor.plugins:
             plugin.preprocess(self, metadata)
@@ -179,6 +181,10 @@ class Post(object):
         return r
 
     @property
+    def content_finalized(self):
+        return self._content_finalized
+
+    @property
     def content_raw(self):
         return self._content_raw
 
@@ -267,6 +273,35 @@ class Post(object):
                                                                                older_post=older_post,
                                                                                all_posts=all_posts,
                                                                                nav_context='post')
+
+    def set_finalized_content(self, content, caller_class):
+        """
+        Plugins can call this method to modify post content that is written back to source post files.
+        This method can be called at any time by anyone, but it has no effect if the caller is not granted the
+        ``MODIFY_RAW_POST`` permission in the Engineer configuration.
+
+        The :attr:`~engineer.conf.EngineerConfiguration.FINALIZE_METADATA` setting must also be enabled in order for
+        calls to this method to have any effect.
+
+        :param content: The modified post content that should be written back to the post source file.
+        :param caller_class: The class of the plugin that's calling this method.
+        :return: ``True`` if the content was successfully modified; otherwise ``False``.
+        """
+        caller = caller_class.get_name() if hasattr(caller_class, 'get_name') else unicode(caller_class)
+        if not settings.FINALIZE_METADATA:
+            logger.warning("A plugin is trying to modify the post content but the FINALIZE_METADATA setting is "
+                           "disabled. This setting must be enabled for plugins to modify post content. "
+                           "Plugin: %s" % caller)
+            return False
+        perms = settings.PLUGIN_PERMISSIONS['MODIFY_RAW_POST']
+        if caller not in perms and '*' not in perms:
+            logger.warning("A plugin is trying to modify the post content but does not have the "
+                           "MODIFY_RAW_POST permission. Plugin: %s" % caller)
+            return False
+        else:
+            logger.debug("%s is setting post source content." % caller)
+            self._content_finalized = content
+            return True
 
     def __unicode__(self):
         return self.slug